Create MeasurementClass.py
A class that allows the creation of "Measurement" objects that can be used for creating recipes, combining ingredients with compatible units. Allows for metric and imperial systems, volume and mass as well as the creation of generic units. By default, adding and subtracting will convert to the highest whole unit of measure when possible.